Make changes after review
